+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixes
+
+-   Fixed the error message shown if a network error occured while consuming a link in passwordless
+
 ## [0.26.4] - 2022-09-29
 
 ### Fixes